Перенес класс List в отдельный файл List.cs
